+++ vitest.config.js
@@ -9,6 +9,11 @@ export default defineConfig({
   plugins: [react()],
   test: {
     environment: 'jsdom',
+    environmentOptions: {
+      jsdom: {
+        url: 'http://localhost/',
+      },
+    },
     globals: true,
     setupFiles: ['./vitest.setup.js'],
     include: ['src/**/*.vitest.{js,jsx,ts,tsx}'],